+++ satella/coding/structures/__init__.py
@@ -4,7 +4,8 @@ from .dictionaries import DictObject, apply_dict_object, DictionaryView, TwoWayD
 from .hashable_objects import HashableWrapper
 from .heaps import Heap, SetHeap, TimeBasedHeap, TimeBasedSetHeap
 from .immutable import Immutable, frozendict
-from .mixins import OmniHashableMixin, ReprableMixin, StrEqHashableMixin, ComparableIntEnum
+from .mixins import OmniHashableMixin, ReprableMixin, StrEqHashableMixin, ComparableIntEnum, \
+    HashableIntEnum
 from .proxy import Proxy
 from .ranking import Ranking
 from .singleton import Singleton, SingletonWithRegardsTo
@@ -18,6 +19,7 @@ __all__ = [
     'ReprableMixin',
     'StrEqHashableMixin',
     'ComparableIntEnum',
+    'HashableIntEnum',
     'DirtyDict',
     'SortedList',
     'SelfCleaningDefaultDict',