+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 # v2.8.18
 
+* add `HashableIntEnum`
+* made `ComparableIntEnum` inherit from `HashableIntEnum`
+* put those two in the docs